fortran: do not link Fortran libs with unnecessary libraries
Fortran libraries do not need to be linked with PMIx, hwloc nor
libevent. That can cause some issues with incompatible flags are pulled
from the .la files.
For example, when the NAG Fortran compiler is used,
libmpi_mpifh_sizeof.la pulls the -pthread flag, that is translated
into -Wl,-pthread.
Then when libmpi_mpifh.la is linked with gcc but pulls -Wl,-pthread that
causes the build to fail since -pthread is not a valid ld option.
Not depending on these libraries is enough to avoid the above issue.

Refs #12413

Thanks to Matthew Thompson for reporting this issue and helping with
the diagnostic.

Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>